backend/check_db.py
import mysql.connector
from database import db_config

def check_columns():
    try:
        conn = mysql.connector.connect(**db_config)
        cursor = conn.cursor()
        cursor.execute("DESCRIBE couriers")
        columns = cursor.fetchall()
        for col in columns:
            print(col)
        conn.close()
    except Exception as e:
        print(f"Error: {e}")

if __name__ == "__main__":
    check_columns()
